Added a readme.md specific to this project
So I realized that I needed some good instruction on github
for how to use this project but that means that I needed to
use the readme.md. And that was really supposed to be blank
for the template but that's not really the case anymore. Anyways
I used it for this project and updated the script to create
an empty one when you create a new project